requirements.txt
streamlit
pandas
plotly
openpyxl
matplotlib
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
streamlit_autorefresh

